Refactor : [#issueNumber-15] 고정확장자 삭제로직 백엔드 코드 개선
- 고정 확장자 삭제 시 uncheck된 상태에서만 삭제가능하도록 개선 (백엔드)

+++ src/main/java/com/flowhyemin/extensionfilter/domain/fixextension/service/FixExtensionService.java
@@ -42,6 +42,9 @@ public class FixExtensionService {
     public FixExtensionDeleteResponse deleteFixExtension(FixExtensionDeleteRequest fixExtensionDeleteRequest) {
         FixExtension fixExtension = fixExtensionRepository.findByName(fixExtensionDeleteRequest.getName())
             .orElseThrow(() -> new FixExtensionException(ErrorCode.NONE_EXISTENCE_FIX_EXTENSION));
+        if (Boolean.TRUE.equals(fixExtension.getIsChecked())) {
+            throw new FixExtensionException(ErrorCode.CHECKED_FIX_EXTENSION);
+        }
         fixExtensionRepository.delete(fixExtension);
         return FixExtensionDeleteResponse.fromEntity(fixExtension);
     }
